# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 権田原佃煮店 外部連携・公式Web API仕様書 (v1)
  description: |
    創業明治38年「権田原佃煮店」の基幹商品マスタ照会、在庫照会、代理受発注、およびレビュー投稿連携を行う公式RESTful API仕様書です。
    提携ECモール・卸売業者・外部システムとの自動連携を想定して標準化されています。
    
    ### 認証方式
    データ更新系API（注文作成等）では、リクエストヘッダーにAPIキーを付与してください。
    * ヘッダー名: `X-Gondawara-Api-Key`
    * テスト用APIキー: `gondawara-secret-api-key-2026`
  version: 1.2.0-STABLE
  contact:
    name: 権田原水産 システム開発推進室
    email: api-support@gondawara-tsukudani.example.com
servers:
  - url: http://localhost:9080/api
    description: ローカル検証環境 (Port 9080)
  - url: https://api.gondawara-tsukudani.example.com/api
    description: ステージング環境
paths:
...
  /v1/orders:
    post:
      summary: 外部代理注文の登録
      description: 提携先モールや卸先からの代理注文を一括登録し、在庫を引き当てます。
      operationId: createOrder
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - item_id
                - quantity
                - customer_name
                - postal_code
                - address
                - phone
              properties:
                item_id:
                  type: integer
                  description: 注文対象の商品ID
                  example: 1
                quantity:
                  type: integer
                  minimum: 1
                  maximum: 10
                  description: 注文数量（1〜10個）
                  example: 2
                customer_name:
                  type: string
                  maxLength: 50
                  description: お届け先氏名
                  example: "権田原 太郎"
                postal_code:
                  type: string
                  pattern: '^\d{3}-\d{4}$'
                  description: 郵便番号（ハイフン必須、7桁）
                  example: "104-0045"
                address:
                  type: string
                  maxLength: 100
                  description: お届け先住所
                  example: "東京都中央区築地4-12-1"
                phone:
                  type: string
                  description: 日中連絡可能な電話番号（ハイフン任意）
                  example: "090-1234-5678"
                notes:
                  type: string
                  maxLength: 200
                  description: 熨斗（のし）等のご要望
                  example: "お中元のし希望（名入れ：権田原）"
      responses:
        '201':
          description: 注文受付・在庫引当完了
          content:
            application/json:
              schema:
                type: object
                required:
                  - order_id
                  - order_number
                  - total_price
                  - status
                properties:
                  order_id:
                    type: integer
                    example: 1003
                  order_number:
                    type: string
                    example: "API-20260927-ABCDE"
                  total_price:
                    type: integer
                    description: 税込合計請求額
                    example: 1936
                  status:
                    type: string
                    example: "API受付完了"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/coupons/verify:
    post:
      summary: 合言葉・クーポンコードの有効性検証
      description: 顧客が入力した常連合言葉やキャンペーンコードを照合し、適用可否および値引き金額を返却します。
      operationId: verifyCoupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                  description: クーポンコード（半角英数字）
                  example: "HIMITSU_50"
      responses:
        '200':
          description: クーポンの判定完了（有効/無効問わず判定結果を返却）
          content:
            application/json:
              schema:
                type: object
                required:
                  - valid
                properties:
                  valid:
                    type: boolean
                    description: クーポンが有効な場合は true、無効な場合は false
                    example: true
                  discount_amount:
                    type: integer
                    description: 固定値引き額（円）
                    example: 500
                  message:
                    type: string
                    example: "常連合言葉が正常に認証されました"
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Gondawara-Api-Key
      description: 外部提携用APIシークレットキー
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "エラーメッセージの内容"
  responses:
    BadRequest:
      description: 不正なリクエストパラメータ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "リクエストパラメータが正しくありません"
    Unauthorized:
      description: 認証エラー (APIキー不正または未指定)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "APIキーが無効または未指定です。"
...
    UnprocessableEntity:
      description: 入力バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "必須パラメータが不足しています。"